Update createMobilityFile.R
New file with Google Mobility Reports up to 15/10/2022 for all UK uploaded to Azure inside `nationaldata-v2`. Changes to the R script to generate it. Note the change to the name of the area column from `CTY20` to `GoogleCTY_CNC` as County only partially reflects the names of the sub-regions at UK level.